data_handling-TST,BUG-test_get_wave_files-> Change the directory the the correct one ('dir' instead of 'path_to_assets'). Check if dir exists before trying to delete it
Former-commit-id: c4f661950dd7bf3bcd6524d3418552ee0e175af6
Former-commit-id: 4f10c32f4327b0750f3c73fce0b9463faff576bd

+++ sound_classification/tests/test_data_handling.py
@@ -165,7 +165,8 @@ def test_to1hot_works_when_when_applying_to_DataFrame(input,depth, expected):
 def test_get_wave_files():
     dir = os.path.join(path_to_assets,'test_get_wave_files')
     #delete directory and files within
-    shutil.rmtree(dir)
+    if os.path.exists(dir):
+        shutil.rmtree(dir)
     dh.create_dir(dir)
     # create two wave files
     f1 = os.path.join(dir, "f1.wav")
@@ -177,7 +178,7 @@ def test_get_wave_files():
     assert len(files) == 2
     assert files[0] == "f1.wav"
     assert files[1] == "f2.wav"
-    files = dh.get_wave_files(path_to_assets, fullpath=True)
+    files = dh.get_wave_files(dir, fullpath=True)
     assert len(files) == 2
     assert files[0] == f1
     assert files[1] == f2
